backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.1",
    "faster-whisper>=1.2.1",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "uvicorn>=0.40.0",
]
